int32 type

+++ my_mrcnn/model.py
@@ -2922,7 +2922,7 @@ def norm_boxes_graph(boxes, shape):
     """
     h, w = tf.split(tf.cast(shape, tf.float32), 2)
     scale = tf.concat([h, w, h, w], axis=-1) - tf.constant(1.0)
-    shift = tf.constant([0., 0., 1., 1.])
+    shift = tf.constant([0., 0., 1., 1.], dtype="int32")
     return tf.divide(boxes - shift, scale)
